feat(bill-manager): Update cached dependency and refactor client code
* Update cached dependency to version 0.52
* Refactor client code to conditionally include openssl-related imports based on feature flags
* Move Invoice and InvoiceItem structs to bill_manager module
