+++ src/lib.rs
@@ -10,8 +10,7 @@ pub mod validator;
 
 pub use client::Mpesa;
 pub use constants::{
-    CommandId, IdentifierTypes, Invoice, InvoiceItem, ResponseType, SendRemindersTypes,
-    TransactionType,
+    CommandId, IdentifierTypes, ResponseType, SendRemindersTypes, TransactionType,
 };
 pub use environment::ApiEnvironment;
 pub use environment::Environment::{self, Production, Sandbox};
