+++ src/services/bill_manager/bulk_invoice.rs
@@ -2,8 +2,8 @@
 
 use serde::Deserialize;
 
+use super::Invoice;
 use crate::client::Mpesa;
-use crate::constants::Invoice;
 use crate::errors::{MpesaError, MpesaResult};
 
 const BILL_MANAGER_BULK_INVOICE_API_URL: &str = "v1/billmanager-invoice/bulk-invoicing";
