+++ tests/mpesa-rust/bill_manager_test/single_invoice_test.rs
@@ -1,5 +1,6 @@
 use chrono::prelude::Utc;
-use mpesa::{InvoiceItem, MpesaError};
+use mpesa::services::InvoiceItem;
+use mpesa::MpesaError;
 use serde_json::json;
 use wiremock::matchers::{method, path};
 use wiremock::{Mock, ResponseTemplate};
